refactor: clean up DDIConverter
